# Server/MagicStreamMoviesServer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for MagicStreamMovies application.",
        "title": "MagicStreamMovies API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/genres": {
            "get": {
                "description": "Get a list of all unique genres",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get all genres",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Genre"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/movie": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new movie to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Add a new movie",
                "parameters": [
                    {
                        "description": "Movie Data",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Movie"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/movie/{imdb_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific movie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get a movie by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IMDB ID",
                        "name": "imdb_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Movie"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/movie/{imdb_id}/review": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the admin review for a movie (requires ADMIN role)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Update admin review (Admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IMDB ID",
                        "name": "imdb_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Admin Review JSON {\\",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/movies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all movies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get all movies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Movie"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/recommendedMovies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get recommended movies based on user's favorite genres",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get recommended movies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Movie"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Genre": {
            "type": "object",
            "required": [
                "genre_id",
                "genre_name"
            ],
            "properties": {
                "genre_id": {
                    "type": "integer"
                },
                "genre_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
        "github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Movie": {
            "type": "object",
            "required": [
                "genre",
                "imdb_id",
                "poster_path",
                "ranking",
                "title",
                "youtube_id"
            ],
            "properties": {
                "admin_review": {
                    "type": "string"
                },
                "genre": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Genre"
                    }
                },
                "id": {
                    "type": "string"
                },
                "imdb_id": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                },
                "ranking": {
                    "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Ranking"
                },
                "title": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 2
                },
                "youtube_id": {
                    "type": "string"
                }
            }
        },
        "github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Ranking": {
            "type": "object",
            "required": [
                "ranking_name",
                "ranking_value"
            ],
            "properties": {
                "ranking_name": {
                    "type": "string"
                },
                "ranking_value": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your Bearer token (e.g., \"eyJhbGci...\"). The \"Bearer \" prefix is optional.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}